move grader to same file
